pyproject.toml
[project]
name = "py-exegia"
version = "0.1.11"
description = "Graph-based biblical and religious text study platform"
readme = "README.md"
license = { text = "Proprietary" }
repository = { url = "https://github.com/Mannydefreitas7/exegia-api-py" }
authors = [{ name = "Emmanuel De Freitas", email = "manny.defreitas7@me.com" }]
requires-python = ">=3.13"
keywords = ["exegia", "bible", "corpus", "text-fabric", "context-fabric", "mcp", "graphql"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # GraphQL
    "strawberry-graphql[fastapi]>=0.314.0",
    # MCP server
    "fastmcp>=2.0",
    # Corpus
    "context-fabric>=0.5.0",
    "text-fabric>=13.0.0",
    # HTTP client
    "httpx>=0.27.0",
    # Config / env
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "dotenvx>=0.0.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyOpenSSL>=23.0.0",
    # Utilities
    "python-multipart>=0.0.9",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.scripts]
cf-mcp = "exegia.mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/exegia"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
